Basic update APIs

+++ app/Model/User.php
@@ -148,7 +148,7 @@ class User extends AppModel
      */
     protected $modelConversionTable = [
         'default_team_id' => DataType::INT,
-        'timezone' => DataType::INT
+        'timezone' => DataType::FLOAT
     ];
 
     /**
